+++ CHANGELOG.md
@@ -4,20 +4,41 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
-
 ## [Unreleased](https://github.com/USGS-WiM/Thresholds/tree/dev)
 
 ### Added
 
--
+- 
 
 ### Changed
 
--
+- 
+
+### Fixed
+
+- 
+
+
+## [v0.7.0](https://github.com/USGS-WiM/Thresholds/releases/tag/v0.7.0) - 9/29/2021
+
+### Added
+
+- Added "Error loading NFHL Services" alert if layer load time > 30 seconds or another error occurs
+- Added Link to "Subscribe to Water Alert" in reference point popups
+
+### Changed
+
+- Set min date to 01/01/2021 for date picker
+- Added time to nwis line on graphs
+- Replaced 'Thresholds' and 'Reference Point' language with 'Feature(s)'
+- Set datepicker to month view for easier date selection
+- Change name of datepicker to "Date"
+- Made datepicker menu open on calendar icon click
 
 ### Fixed
 
 - Fixed bug in thresholds exceeded count after date change
+- Fix nfhl loading alert bug when unchecking nfhl layer before layer is loaded
 
 ## [v0.6.0](https://github.com/USGS-WiM/Thresholds/releases/tag/v0.6.0) - 9/22/2021